more _runtime cleanup
