+++ website/src/routes/_benchmarks/json-schema/-components/card.tsx
@@ -9,7 +9,7 @@ import {
   jsonSchemaSourceProps,
   standardJsonSchemaProps,
   jsonSchemaTargetProps,
-} from "#src/routes/_benchmarks/_runtime/-constants";
+} from "#src/routes/_benchmarks/json-schema/-constants";
 import { ChipCollection, DisplayChip } from "#src/shared/components/chip";
 import { CodeBlock } from "#src/shared/components/code";
 import { MdSymbol } from "#src/shared/components/symbol";
