+++ website/src/routes/_benchmarks/json-schema/-components/table.tsx
@@ -12,7 +12,7 @@ import {
   jsonSchemaSourceProps,
   standardJsonSchemaProps,
   jsonSchemaTargetProps,
-} from "#src/routes/_benchmarks/_runtime/-constants";
+} from "#src/routes/_benchmarks/json-schema/-constants";
 import { ToggleButton } from "#src/shared/components/button/toggle";
 import { Radio } from "#src/shared/components/radio";
 import { Scaler } from "#src/shared/components/scaler";
