+++ website/src/routes/_home/content.mdx
@@ -3,6 +3,7 @@ import { BenchPlot } from "#src/routes/_benchmarks/_runtime/-components/plot";
 import { DownloadPlot } from "#src/routes/_benchmarks/download/-components/plot";
 import { Admonition } from "#src/shared/components/admonition";
 import { CodecPlot } from "#src/routes/_benchmarks/_runtime/codec/-components/plot";
+import { JsonSchemaPlot } from "#src/routes/_benchmarks/json-schema/-components/plot";
 
 Welcome to the Schema Benchmarks project. This aims to compare the performance of different schema validation libraries in detail, including separating each step of the process.
 
@@ -164,11 +165,11 @@ Some report `0`: `ajv`, `ata-validator` and `typebox` schemas _are_ JSON Schema,
 
 Generating a draft 2020-12 JSON schema of the input type:
 
-<BenchPlot type="jsonSchema" target="draft-2020-12" direction="input" />
+<JsonSchemaPlot target="draft-2020-12" direction="input" />
 
 Generating a draft 2020-12 JSON schema of the output type:
 
-<BenchPlot type="jsonSchema" target="draft-2020-12" direction="output" />
+<JsonSchemaPlot target="draft-2020-12" direction="output" />
 
 ### <Link to="/codec">Codec</Link>
