+++ website/vite.config.ts
@@ -17,15 +17,17 @@ import { defineConfig } from "vitest/config";
 import {
   dataTypeProps,
   errorTypeProps,
-  jsonSchemaDirectionProps,
-  jsonSchemaSourceProps,
-  standardJsonSchemaProps,
-  jsonSchemaTargetProps,
   optimizeTypeProps,
   stringFormatProps,
 } from "#src/routes/_benchmarks/_runtime/-constants";
 import { minifyTypeProps } from "#src/routes/_benchmarks/download/-constants";
 import { speedPresets } from "#src/routes/_benchmarks/download/-speed";
+import {
+  jsonSchemaDirectionProps,
+  jsonSchemaSourceProps,
+  standardJsonSchemaProps,
+  jsonSchemaTargetProps,
+} from "#src/routes/_benchmarks/json-schema/-constants";
 import { admonitionDefaults } from "#src/shared/components/admonition/constants";
 import { sidebarGroups } from "#src/shared/components/sidebar/groups";
 import { sortDirectionIcons } from "#src/shared/components/table/constants";
